added sub-command clean (#421450)
--HG--
extra : convert_revision : schettino72%40gmail.com-20091002111920-wgyz6doo1rkj8ru5

+++ tests/test_dependency.py
@@ -27,8 +27,7 @@ TESTDB = "testdb"
 
 class DependencyTestBase(object):
     def setUp(self):
-        if os.path.exists(TESTDB):
-            os.remove(TESTDB)
+        if os.path.exists(TESTDB): os.remove(TESTDB)
         self.d = Dependency(TESTDB)
 
     def tearDown(self):
